requirements.txt
torch>=2.2.0
pypdf>=4.2.0
numpy>=2.2.0
pymupdf>=1.27.0
pytesseract>=0.3.13
pillow>=12.0.0
fastapi>=0.116.0
uvicorn>=0.35.0
httpx>=0.28.0
